Make it possible to set the planning scene used.
Collision checking will now use the robot state from the planning scene as the basis (in particular, attached collision objects here will be respected).
Collision checking is no longer only done for the group used for planning.

A bit more work is probably still needed here. Also would be nice to get rid of some of the compiler warnings while we are working on this anyways.